Complete bibtex support
